Replaced all variable length arrays with std::vector's, and "long long" with "long", to comply with the ISO C++ 1998 standard, which is enforced by CRAN.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/bfp@30 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/types.h
@@ -31,9 +31,7 @@ typedef int Int;
 typedef unsigned int PosInt;
 
 // general large unsigned integer
-typedef unsigned long long int PosLargeInt;
-
-
+typedef unsigned long int PosLargeInt;
 
 // here is an integer vector
 typedef std::vector<Int> IntVector;
